generate_pattern_map and helper functions added
This function is used to precalculate the patterns from a dictionary and to store it in chunks that should make subsequent querying and processing much faster and more efficient.

build_word_pattern can also be used by the main product to calculate a word's pattern and unique characters count so that querying is possible.